Fix T93154: Crash adding multiple movie strips
Some when adding multiple movies at once and only some of them have
audio track, this causes crash on NULL dereference. Issue was introduced
in bdbc7e12a02e to align sound and video properly.

Check if sound is present in movie file. If it's not, don't try to align
sound with video.